requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.7.2
xgboost==3.0.5
openpyxl==3.1.5
xlwt==1.3.0
scipy==1.16.2
joblib==1.5.2
threadpoolctl==3.6.0
supabase==2.3.4
python-dotenv==1.0.0
requests==2.31.0

# PDF extraction helpers (the script tries tabula first, then camelot, then pdfplumber)
tabula-py==2.9.0
camelot-py[cv]==0.11.0
pdfplumber==0.11.0
PyPDF2==3.0.1
